refactor(isInteropObservable): rename Observablelike and isObservable
- ObservableLike deprecated in favor of InteropObservable (as a name)
- renamed isObservable to isInteropObservable
