+++ spec/observables/from-spec.ts
@@ -34,7 +34,7 @@ describe('from', () => {
     expect(r).to.throw();
   });
 
-  type('should return T for ObservableLike objects', () => {
+  type('should return T for InteropObservable objects', () => {
     /* tslint:disable:no-unused-variable */
     const o1: Observable<number> = from([] as number[], asapScheduler);
     const o2: Observable<{ a: string }> = from(Observable.empty());
